requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
blosc2==3.9.1
Brotli==1.1.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
cityscapesScripts==2.2.4
click==8.3.0
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
descartes==1.1.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@fd27788985af0f4ca800bca563acdb700bb890e2
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.118.0
ffmpy==0.6.1
filelock==3.19.1
fire==0.7.1
fonttools==4.60.1
fsspec==2025.9.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.45
gradio==5.49.0
gradio_client==1.13.3
groovy==0.1.2
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.34.0
iopath==0.1.9
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
KNN==0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
ndindex==1.10.0
networkx==3.4.2
numexpr==2.13.1
numpy==1.26.4
nuscenes-devkit==1.2.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open_clip_torch==2.31.0
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
orjson==3.11.3
packaging==25.0
pandas==2.3.3
parameterized==0.9.0
peft==0.10.0
pillow==10.2.0
platformdirs==4.4.0
portalocker==3.2.0
protobuf==6.32.1
py-cpuinfo==9.0.0
pycocotools==2.0.10
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
RandomPatchExtraction==0.1
requests==2.32.5
rich==14.1.0
ruff==0.13.3
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.12.0
semantic-version==2.10.0
sentry-sdk==2.45.0
setproctitle==1.3.7
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.48.0
sympy==1.13.1
tables==3.10.1
tabulate==0.9.0
termcolor==3.1.0
threadpoolctl==3.6.0
timm==1.0.15
tomlkit==0.13.3
torch==2.5.1
torchaudio==2.8.0
torchvision==0.23.0+cu126
tqdm==4.67.1
trimesh==4.8.3
triton==3.1.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
umpy==0.1.0
urllib3==2.5.0
uvicorn==0.37.0
wandb==0.16.5
websockets==15.0.1
xformers==0.0.28.post3
yacs==0.1.8
